# apps/agent/src/lotto_agent/schemas_generated.py
# generated by datamodel-codegen:
#   filename:  openapi.json
#   timestamp: 2026-06-18T00:07:53+00:00

from __future__ import annotations

from enum import Enum
from typing import Any

from pydantic import BaseModel, Field, RootModel, conint, model_validator


class LottoCombinationItem(RootModel[conint(ge=1, le=45)]):
    root: conint(ge=1, le=45)


class LottoCombination(RootModel[list[LottoCombinationItem]]):
    root: list[LottoCombinationItem] = Field(..., max_length=6, min_length=6)

    @model_validator(mode='after')
    def validate_ascending_unique(self):
        numbers = [int(item.root) for item in self.root]
        if numbers != sorted(set(numbers)):
            raise ValueError('lotto combination must be unique and ascending')
        return self


class LuckyNumber(RootModel[conint(ge=1, le=45)]):
    root: conint(ge=1, le=45)


class RecommendationRequest(BaseModel):
    luckyNumbers: list[LuckyNumber] = Field(..., max_length=6)
    count: conint(ge=1, le=50)
    targetDrawNo: conint(ge=1) | None = None


class Status(Enum):
    OK = 'OK'
    PARTIAL = 'PARTIAL'
    INFEASIBLE_LUCKY_SET = 'INFEASIBLE_LUCKY_SET'


class Feasibility(BaseModel):
    status: Status
    requestedCount: int
    effectiveCount: int
    partial: bool
    message: str | None = None


class RecommendationBase(BaseModel):
    id: str
    numbers: LottoCombination
    tags: list[str]
    stats: dict[str, Any]


class Recommendation(RecommendationBase):
    explanation: str
    tagNarration: str


class SyncStatus(Enum):
    fresh = 'fresh'
    last_good = 'last-good'
    skipped = 'skipped'


class DataFreshness(BaseModel):
    latestSyncedDrawNo: conint(ge=0)
    syncStatus: SyncStatus
    syncErrorKind: str | None


class RecommendationResponse(BaseModel):
    requestId: str
    traceId: str | None
    targetDrawNo: int
    luckyNumbers: list[conint(ge=1, le=45)]
    recommendations: list[Recommendation]
    fallbackUsed: bool
    freshness: DataFreshness
    feasibility: Feasibility
    disclaimers: list[str]


class ExplainRequest(BaseModel):
    requestId: str
    traceId: str | None
    combinations: list[RecommendationBase]
    stats: dict[str, Any]
    luckyNumbers: list[conint(ge=1, le=45)]
    targetDrawNo: int


class PerCombinationItem(BaseModel):
    id: str
    explanation: str
    tagNarration: str


class ExplainResponse(BaseModel):
    perCombination: list[PerCombinationItem]
    analysisSummary: str
    fallbackUsed: bool


# Project helper generated from LottoCombination schema invariants.
def _lotto_numbers(value: LottoCombination) -> list[int]:
    return [int(item.root) for item in value.root]


def validate_lotto_combination(value: LottoCombination) -> LottoCombination:
    return LottoCombination.model_validate(value)