# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AICanWinLottery API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/recommendations": {
      "post": {
        "operationId": "createRecommendations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommendation response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/explain": {
      "post": {
        "operationId": "explain",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Explanation response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LottoCombination": {
        "type": "array",
        "minItems": 6,
        "maxItems": 6,
        "uniqueItems": true,
        "x-ascending": true,
        "items": {
          "type": "integer",
          "minimum": 1,
          "maximum": 45
        }
      },
      "RecommendationRequest": {
        "type": "object",
        "required": [
          "luckyNumbers",
          "count"
        ],
        "properties": {
          "luckyNumbers": {
            "type": "array",
            "maxItems": 6,
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 45
            }
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          },
          "targetDrawNo": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "Feasibility": {
        "type": "object",
        "required": [
          "status",
          "requestedCount",
          "effectiveCount",
          "partial"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OK",
              "PARTIAL",
              "INFEASIBLE_LUCKY_SET"
            ]
          },
          "requestedCount": {
            "type": "integer"
          },
          "effectiveCount": {
            "type": "integer"
          },
          "partial": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RecommendationBase": {
        "type": "object",
        "required": [
          "id",
          "numbers",
          "tags",
          "stats"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "numbers": {
            "$ref": "#/components/schemas/LottoCombination"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stats": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "Recommendation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RecommendationBase"
          },
          {
            "type": "object",
            "required": [
              "explanation",
              "tagNarration"
            ],
            "properties": {
              "explanation": {
                "type": "string"
              },
              "tagNarration": {
                "type": "string"
              }
            }
          }
        ]
      },
      "DataFreshness": {
        "type": "object",
        "required": [
          "latestSyncedDrawNo",
          "syncStatus",
          "syncErrorKind"
        ],
        "properties": {
          "latestSyncedDrawNo": {
            "type": "integer",
            "minimum": 0
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "fresh",
              "last-good",
              "skipped"
            ]
          },
          "syncErrorKind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "RecommendationResponse": {
        "type": "object",
        "required": [
          "requestId",
          "traceId",
          "targetDrawNo",
          "luckyNumbers",
          "recommendations",
          "fallbackUsed",
          "freshness",
          "feasibility",
          "disclaimers"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetDrawNo": {
            "type": "integer"
          },
          "luckyNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 45
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            }
          },
          "fallbackUsed": {
            "type": "boolean"
          },
          "freshness": {
            "$ref": "#/components/schemas/DataFreshness"
          },
          "feasibility": {
            "$ref": "#/components/schemas/Feasibility"
          },
          "disclaimers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExplainRequest": {
        "type": "object",
        "required": [
          "requestId",
          "traceId",
          "combinations",
          "stats",
          "luckyNumbers",
          "targetDrawNo"
        ],
        "properties": {
          "requestId": {
            "type": "string"
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "combinations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationBase"
            }
          },
          "stats": {
            "type": "object",
            "additionalProperties": true
          },
          "luckyNumbers": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 45
            }
          },
          "targetDrawNo": {
            "type": "integer"
          }
        }
      },
      "ExplainResponse": {
        "type": "object",
        "required": [
          "perCombination",
          "analysisSummary",
          "fallbackUsed"
        ],
        "properties": {
          "perCombination": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "explanation",
                "tagNarration"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "explanation": {
                  "type": "string"
                },
                "tagNarration": {
                  "type": "string"
                }
              }
            }
          },
          "analysisSummary": {
            "type": "string"
          },
          "fallbackUsed": {
            "type": "boolean"
          }
        }
      }
    }
  }
}